requirements.txt
duckduckgo-search
google-generativeai
requests
beautifulsoup4
flask
Gunicorn